feat: Updates shotgun version to 1.1.0 release (#139)
**Requirements**

- [ ] I have added test coverage for new or changed functionality
- [ ] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [ ] I have validated my changes against all supported platform
versions

**Related issues**

Provide links to any issues in this repository or elsewhere relating to
this pull request.
None found.

**Describe the solution you've provided**
There was a new version of shotgun released on Oct 24 that has an
additional callback for a down message. Upgrading the dependency will
support that.

**Describe alternatives you've considered**
The alternative would be not to update the dependencies

**Additional context**
The applications I am using this in are getting those messages sent now
and ending up in undefined function errors in our error tracking tool.
It is not a great consequence, but upgrading the dependency seems like a
good practice anyway.