feat: Update minimum OTP version to 24 (#156)
BEGIN_COMMIT_OVERRIDE
feat: Update minimum OTP version to 24.
fix: Fix issue where sometimes a dropped stream would not reconnect.
END_COMMIT_OVERRIDE

In shotgun 1.1.0 a new feature was introduced which introduced a bug
with the SSE connection. Where previously it would close its process in
all cases where it disconnected, now it would not do that in some
situations. This caused #155.

We contributed an update to shotgun to allow disabling the feature for
async connections where you need to monitor the connection state.
https://github.com/inaka/shotgun/pull/203

This was released in shotgun 1.2.1.

This PR updates to 1.2.1 and uses the option to disable the `reopen`
feature added in shotgun 1.1.0.
This PR also raises the minimum erlang version to 24. This is required
to gun 2.2 which shotgun 1.2.0 and newer depends on. OTP doesn't have an
official support timeline, but OTP 24 came out in 2021, so most people
will likely be on 24 or greater especially when using elixir.