fix: Handle connection errors during read operations. (#163)
This PR adds error handling for read operations from the redis
persistent store. This is an incremental step in addressing #161.

In the case we cannot read from redis, we will treat that as a flag not
being found. If we cannot read from redis when doing an initialization
check, then we will consider the store not to have been initialized.

This addresses the primary issues with daemon mode, but it doesn't
address persistent store errors in non-daemon mode, or other cases that
could cause the redis process to exit. Subsequent PRs will address these
issues.

The detailed variation interface does not report why a flag cannot be
found, and isn't the correct avenue for doing so, and as such we may
have to consider supporting the data store status interface. This is
also not tackled in this PR.