fix: Redact SDK in gen_server descriptions and network errors. (#166)
There are two categories of issues addressed in this PR.

The first is that there are situations where an SDK key can be logged if
it is part of process state. So this PR introduces format_status
callbacks which control how that formatting is done, and redacts the SDK
key.

The second is there are some network errors that could log the headers
which include authorization. This PR adds some extra formatting to
network errors to reduce the detail contained in the error. A
side-effect will be that in some situations logs will contain less
useful data for troubleshooting.